Reverted some types

+++ src/correlate.cpp
@@ -77,7 +77,7 @@ void normalize_correlation(
 ){
     double cr1 = (double) np1 / (double) dt1;
     double cr2 = (double) np2 / (double) dt2;
-    for(size_t j=0; j < x_axis.size(); j++){
+    for(int j=0; j<x_axis.size(); j++){
         uint64_t pw = (uint64_t) std::pow(2.0, (int) (float(j-1) / n_bins));
         double t_corr = (dt1 < dt2 - x_axis[j]) ? (double) dt1 : (double) (dt2 -
                 x_axis[j]);
