+++ src/image.cpp
@@ -160,7 +160,7 @@ void CLSMImage::initialize(TTTR* tttr_data){
 
     // find the frame and line markers
     for(
-            size_t i=0;
+            unsigned int i=0;
             i < tttr_data->n_valid_events;
             i++)
     {
@@ -188,7 +188,7 @@ void CLSMImage::initialize(TTTR* tttr_data){
                 frame->push_back(line);
 
                 // Find line end
-                for(size_t j=i + 1; j < tttr_data->n_valid_events; j++){
+                for(unsigned j=i + 1; j<tttr_data->n_valid_events; j++){
                     c = tttr_data->routing_channels[j];
                     e = tttr_data->event_types[j];
                     if( (e == marker_event) && (c == marker_line_stop) ){
